requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pywebview==4.4.1
Pillow==10.1.0
PyInstaller==6.3.0 